// frontend/src/core/models/company.model.ts
export interface CompanyResponse {
    symbol: string;
    companyName: string;
    exchange: string;
    industry: string;
    sector: string;
    country: string;
    currency: string;
    website: string;
    description: string;
    marketCap: number;
    sharesOutstanding: number;
    source: string;
    updatedAt: string;
}